requirements.txt
asgiref==3.11.0
Django==5.2.8
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
Pillow==10.4.0
gunicorn==23.0.0
rasterio==1.3.10
mercantile==1.2.1
pyproj==3.7.2
